backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
lighthouseweb3>=0.2.4
web3>=6.0.0
# ML dependencies
pandas>=1.5.0
scikit-learn>=1.2.0
joblib>=1.2.0
requests>=2.28.0
# ML dependencies will be added later
# pandas
# scikit-learn
# joblib 